Reformatted builtins
