+++ src/shell.c
@@ -64,7 +64,7 @@ void shell_print_line_flair(void)
 
 void shell_print_greeting(void)
 {
-    printf("Welcome to "BLU "asn" CRESET " shell.\n\nInput "BLU "help"CRESET " for a list of commands.\n");
+    printf("Welcome to "BLU "asn" CRESET " shell.\n\nInput " BLU "help" CRESET " for a list of commands.\n");
 }
 
 void shell_loop_step(bool print_output, bool print_input)
